Cancel Dvs create order, get dvsId from Optimus, Get from previous process optimus taskId for every process, New field send in optimus create process

+++ src/main/java/az/kapitalbank/marketplace/client/optimus/model/process/ProcessData.java
@@ -13,8 +13,8 @@ import static lombok.AccessLevel.PRIVATE;
 @FieldDefaults(level = PRIVATE)
 public class ProcessData {
     String cif;
-    int period;
-    int dvsOrderId;
+    Integer period;
+    Long dvsOrderId;
     String salaryPan;
     MainIncome mainIncome;
     SelectedOffer selectedOffer;
